requirements.txt
argcomplete==3.5.1
asgiref==3.8.1
Beaker==1.12.1
beautifulsoup4==4.12.3
blivet==3.10.0
blivet-gui==2.5.0
boto3==1.35.46
botocore==1.35.46
Brlapi==0.8.5
cffi==1.16.0
charset-normalizer==3.3.2
click==7.1.2
cryptography==41.0.7
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
distlib==0.3.8
distro==1.9.0
Django==4.2.15
djangorestframework==3.15.2
dnf==4.21.1
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.15.4
Flask==1.1.4
fros==1.1
gpg==1.23.2
graphviz==0.20.3
humanize==3.13.1
idna==3.7
iso639==0.1.4
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==1.0.1
langtable==0.0.68
libcomps==0.1.20
libdnf==0.73.3
louis==3.28.0
lxml==5.1.0
Mako==1.2.3
MarkupSafe==2.0.1
nftables==0.1
olefile==0.47
packaging==23.2
Paste==3.7.1
pexpect==4.9.0
pid==2.2.3
pillow==10.3.0
platformdirs==4.2.2
ply==3.11
productmd==1.38
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pwquality==1.4.5
py-postgresql==1.2.2
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pyenchant==3.2.2
PyGObject==3.48.2
pykickstart==3.52
pyOpenSSL==23.2.0
pyparted==3.13.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-meh==0.51
python-pam==2.0.2
pyudev==0.24.1
pyxdg==0.27
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
requests-file==2.0.0
requests-ftp==0.3.1
rpm==4.19.1.1
s3transfer==0.10.3
selinux @ file:///builddir/build/BUILD/libselinux-3.7/src
sepolicy @ file:///builddir/build/BUILD/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==69.0.3
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.7.2
soupsieve==2.5
sqlparse==0.5.1
systemd-python==235
Tempita==0.5.2
torbrowser-launcher==0.3.7
typing_extensions==4.12.2
urllib3==1.26.20
virtualenv==20.26.3
Werkzeug==1.0.1
xkbregistry==0.3
